Update README.md
Here is a code please help me to attached a .html Form with this code to start calculating employees salary. I am new on this platform and want to know coding,my heartiest request to all of you Please consider me as student of you. any suggestions,idea and commands are always acceptable.